+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
     - Sanitize and validate ``IRODS_ROOT_PATH`` in ``get_root_path()`` (#1891)
 - **Landingzones**
     - Create assay plugin shortcut collections for zones (#1869)
+    - Zone statistics for siteinfo (#1898)
 - **Taskflowbackend**
     - Django settings in siteinfo (#1901)